Fix Number Bar properties dialog not showing (#634)
Fix Number Bar properties dialog not showing when "Edit properties" is
selected in the context menu (#634). This is caused by NumberBarWidget
incorrectly using a DoubleProperty instead of an IntegerProperty for the
numTickMarks property, causing a ClassCastException when trying to cast
from Number to Integer when displaying the property in a dialog.

Increase the version of the base plugin from 1.1.5 to 1.1.6.