Remove large cache of all data points
The realData map used a PrimitiveDoubleArrayList in order to store all the numbers. This was used to allow the user to arbitrarily change the visibleTime property and have past data reemerge. This however, created a large memory leak, so the feature has been removed. PrimitiveDoubleArrayList has also been removed as it is no longer needed.