Fixes for NT-backed widgets when the server restarts
Widgets were getting destroyed sources added, but weren't replacing the backing NT source. These widgets would be permanently disabled and would need to be deleted and recreated, or a dashboard application restart, to be usable again

Topics written to by the dashboard would not receive "unpublish" events from the server, so they'd always be active and available. The `allUris` observable list was incorrectly adding/removing entries so URIs for those topics could still be removed from the list (disabling connected widgets with a destroyed source) if those URIs happened to be at the same index in the `allUris` list as a removed item in the plugin's URI list

Topics written to by the dashboard would not receive a "publish" event from the server on reconnect, so no data sources would be created for them. We now manually trigger the data source creation logic for every topic that still exists in the NT client upon a server reconnect to ensure those sources are recreated